+++ test/mock/hit/pageview.js
@@ -1,5 +1,5 @@
 export default {
-  type: 'Page',
+  type: 'PageView',
   data: {
     documentLocation: 'http%3A%2F%2Fabtastylab.com%2F60511af14f5e48764b83d36ddb8ece5a%2F',
     pageTitle: 'YoloPage',