+++ RELEASENOTES.md
@@ -1,5 +1,15 @@
 # Flagship JS SDK - Release notes
 
+## ➡️ Version 2.1.11
+
+- Adds Pageview hit
+- Changes Screen hit to Screenview hit
+- Minor bug fixes
+
+## ➡️ Version 2.1.10
+
+- Bumps axios version from 0.19.2 to 0.21.1
+
 ## ➡️ Version 2.1.9
 
 -   Minor changes.